Move mission.js into pages/missions
