+++ src/components/pages/missions/mission.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { loadMissions } from '../../redux/missions/missions';
-import MissionsTable from './missions/missionsTable';
-import '../css/missions.scss';
+import { loadMissions } from '../../../redux/missions/missions';
+import MissionsTable from './missionsTable';
+import '../../css/missions.scss';
 
 const mission = () => {
   const missionsList = useSelector(({ missionsReducer }) => missionsReducer);
